+++ src/components/Compose/ComposeImage.tsx
@@ -3,7 +3,7 @@ import { memo } from 'react';
 import { RemoveButton } from '@/components/RemoveButton.js';
 import { Image } from '@/esm/Image.js';
 import { classNames } from '@/helpers/classNames.js';
-import { resolveMediaObjectPreviewUrl } from '@/helpers/resolveMediaObjectPreviewUrl.js';
+import { resolveMediaObjectUrl } from '@/helpers/resolveMediaObjectUrl.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
 import type { MediaObject } from '@/types/compose.js';
 
@@ -28,7 +28,7 @@ export const ComposeImage = memo(function ComposeImage({ index, size, image, rea
                 'h-auto': size >= 5,
             })}
         >
-            <Image src={resolveMediaObjectPreviewUrl(image)} alt={image.file.name} fill className="object-cover" />
+            <Image src={resolveMediaObjectUrl(image)} alt={image.file.name} fill className="object-cover" />
 
             {!readonly ? (
                 <RemoveButton className="absolute right-1 top-1 z-10" onClick={() => removeImage(image)} />
