+++ src/helpers/createDummyPost.ts
@@ -5,7 +5,7 @@ import type { SocialSource } from '@/constants/enum.js';
 import { readChars } from '@/helpers/chars.js';
 import { createDummyProfile } from '@/helpers/createDummyProfile.js';
 import { getCurrentProfileAll } from '@/helpers/getCurrentProfile.js';
-import { resolveMediaObjectPreviewUrl } from '@/helpers/resolveMediaObjectPreviewUrl.js';
+import { resolveMediaObjectUrl } from '@/helpers/resolveMediaObjectUrl.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 import type { CompositePost } from '@/store/useComposeStore.js';
 
@@ -41,7 +41,7 @@ export function createDummyCommentPost(source: SocialSource, compositePost: Comp
         mediaObjects: compact([compositePost.video, ...compositePost.images, ...compositePost.images]).map((x) => ({
             title: x.file.name,
             mimeType: x.mimeType,
-            url: resolveMediaObjectPreviewUrl(x),
+            url: resolveMediaObjectUrl(x),
         })),
         metadata: {
             locale: '',
