+++ src/helpers/downloadMediaObjects.ts
@@ -1,4 +1,4 @@
-import { resolveMediaObjectPreviewUrl } from '@/helpers/resolveMediaObjectPreviewUrl.js';
+import { resolveMediaObjectUrl } from '@/helpers/resolveMediaObjectUrl.js';
 import { type MediaObject, MediaSource } from '@/types/compose.js';
 
 async function downloadUrl(url: string, name: string) {
@@ -9,7 +9,7 @@ async function downloadUrl(url: string, name: string) {
 export async function downloadMediaObjects(medias: MediaObject[]) {
     return await Promise.all(
         medias.map(async (media) => {
-            const url = resolveMediaObjectPreviewUrl(media, [MediaSource.Giphy]);
+            const url = resolveMediaObjectUrl(media, [MediaSource.Giphy]);
             return {
                 ...media,
                 file: url ? await downloadUrl(url, media.file.name) : media.file,
